Added skipping to next minima and and lookahead monotonicity is skipped close
if we always insist on the first local maxima in the distance we get caught
on corners with the trajectory lookahead. So if we are close to the start
of the local plan, we wil allow a non-increasing distance to the robot.
And the getNextMinima is allowed to skip short distances, this helps it get
around corners when the vehicle cuts them as long as we got close enough